+++ src/Spy/SpyInterface.ts
@@ -1,0 +1,6 @@
+/**
+ * Common interface for injected spies into tested code.
+ */
+export interface SpyInterface {
+    call(): void;
+}