+++ src/Test/index.ts
@@ -1,2 +1,3 @@
-export {TestCaseMethod} from "./TestCaseMethod";
 export {AbstractResultAdapter} from "./AbstractResultAdapter";
+export {ExpectedExceptionContainer} from "./ExpectedExceptionContainer";
+export {TestCaseMethod} from "./TestCaseMethod";